style(datepicker): split up react index file
